+++ src/common/js/constants.js
@@ -1,0 +1,5 @@
+/*
+ * pwix:modal-info/src/common/js/constants.js
+ */
+
+I18N = 'pwix:modal-info+i18n-internationalization';