+++ src/common/js/global.js
@@ -1,4 +1,5 @@
 /*
  * pwix:modal-info/src/common/js/global.js
  */
+
 ModalInfo = {};